api/requirements.txt
boto3==1.12.19
botocore==1.15.19
click==7.1.1
docutils==0.15.2
Flask==1.0
futures==3.3.0
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-dotenv==0.12.0
s3transfer==0.3.3
six==1.14.0
typing==3.7.4.1
urllib3==1.25.8
Werkzeug==1.0.0
